add_amendment_werewolf_game/requirements.txt
agentscope
pydantic
numpy
python-dotenv
